Refactor classes into their own files and update references
