+++ src/SmugMugShared/Descriptors/Limits.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Alex Ghiondea. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace SmugMug.Shared.Descriptors
+{
+    public class Limits
+    {
+        // INFINITY, NEGATIVE_INFINITY, POSITIVE_INFINITY
+
+        public string Min { get; set; }
+        public string Max { get; set; }
+    }
+}
