+++ src/SmugMugShared/Descriptors/Method.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) Alex Ghiondea. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
+
+namespace SmugMug.Shared.Descriptors
+{
+    public class Method
+    {
+        public string NormalizedUri { get; set; }
+        public string Uri { get; set; }
+        public string ReturnType { get; set; }
+    }
+}
