+++ src/SmugMugShared/Descriptors/Properties/ArrayProperty.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) Alex Ghiondea. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Newtonsoft.Json.Linq;
+using SmugMug.Shared.Extensions;
+
+namespace SmugMug.Shared.Descriptors
+{
+    public class ArrayProperty : Property
+    {
+        public string ItemType { get; set; }
+        public Limits ArrayLimits { get; set; }
+
+        public ArrayProperty(JObject obj)
+            : base(obj)
+        {
+            ItemType = obj.GetValueAsString("ITEM_TYPE");
+
+            var min = obj.Property("MIN_COUNT").ToObject<string>();
+            var max = obj.Property("MAX_COUNT").ToObject<string>();
+            ArrayLimits = new Limits() { Min = min, Max = max };
+        }
+
+        public override string ToString()
+        {
+            return base.ToString() + ": Array (of" + ItemType + ")";
+        }
+    }
+}
