+++ src/SmugMugShared/Descriptors/Properties/DateTimeProperty.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Alex Ghiondea. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Newtonsoft.Json.Linq;
+
+namespace SmugMug.Shared.Descriptors
+{
+    public class DateTimeProperty : Property
+    {
+        public DateTimeProperty(JObject obj)
+            : base(obj)
+        {
+
+        }
+
+        public override string ToString()
+        {
+            return base.ToString() + ": DateTime";
+        }
+    }
+}
