+++ src/SmugMugShared/Descriptors/Properties/DecimalProperty.cs
@@ -1,0 +1,27 @@
+﻿// Copyright (c) Alex Ghiondea. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Newtonsoft.Json.Linq;
+
+namespace SmugMug.Shared.Descriptors
+{
+    public class DecimalProperty : Property
+    {
+        public string Precision { get; set; }
+        public string Scale { get; set; }
+        public bool Signed { get; set; }
+
+        public DecimalProperty(JObject obj)
+            : base(obj)
+        {
+            Precision = obj.Property("PRECISION").ToObject<string>();
+            Scale = obj.Property("SCALE").ToObject<string>();
+            Signed = obj.Property("SIGNED").ToObject<bool>();
+        }
+
+        public override string ToString()
+        {
+            return base.ToString() + string.Format(": Decimal ({0}, {1}), {2}", Precision, Scale, Signed ? "signed" : "unsigned");
+        }
+    }
+}
